making the data type of estimated time to float

+++ estimator/models.py
@@ -71,7 +71,7 @@ class Activity(models.Model):
     feature = models.ForeignKey(Feature, on_delete=models.CASCADE)
     name = models.CharField(max_length=2000, default='')
     estimate = models.ForeignKey(Estimate, on_delete=models.CASCADE)
-    estimated_time = models.IntegerField(default=0)
+    estimated_time = models.FloatField(default=0)
     status = models.CharField(max_length=1, choices=STATUS_CHOICES)
 
     def __str__(self):
@@ -85,7 +85,7 @@ class SubActivity(models.Model):
     STATUS_CHOICES = STATUSES
     name = models.CharField(max_length=2000, default='')
     parent = models.ForeignKey(Activity, on_delete=models.CASCADE)
-    estimated_time = models.IntegerField(default=0)
+    estimated_time = models.FloatField(default=0)
     note = models.TextField(max_length=10000, null=True)
     is_completed = models.BooleanField(default=False)
     status = models.CharField(max_length=1, choices=STATUS_CHOICES)
